[python][wip] add binding generator script
this script parses a file header and generate
function signature and extracts doxygen frames compatible
with pybind11

It was used to generate most of the docs header
